Bring this in line with Easy calculator
